chatbot/requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
flask_sqlalchemy==3.1.1
faiss-cpu==1.8.0
langchain==0.2.4
langchain_community==0.2.4
langchain_huggingface==0.0.3
langchain_openai==0.1.8
python-dotenv==1.0.1
